fix: Makes ts, node12, and node classic imports work
To make everything compile with Typescript, support node12 style exports
in the package.json, and ensure commonjs imports still work for
non-node12 compatible bundlers, we need to export our modules in a way
that exposes a hybrid build with cjs and esm, but also a custom nested
package.json for each module type because of how node manages
subdirectories. Finally, a top level file is exposed for non-node12,
allowing for imports of @taskless/client/next.