refactor: Updates package syncing functionality
This update adds syncpack to the package dependencies and ensures that
all deps are using the same semver compatible ranges. This is mostly for
tools such as GraphQL which need to resolve to a single version, but
can also avoid version inconsistency. As part of this commit, package
files were reformatted to follow a convention.