refactor: Uses zod for type generation and removes graphql dependency
This changeset lightens the total payload of the Taskless client,
replacing the generic type guard with zod. Additionally, the code
generation for the for.taskless.io endpoint was replaced with regular
types. This is because future graphql-code-generator will depend on
the new TypedDocumentNode; awesome for end developers, but terrible for
libraries which should not have a transient dependency on graphql-js